cleanup, review changes, docs
